[dev] feat: update new blog and change how to render the sitemap from dynamic to static and dynamic
