+++ READMe.md
@@ -1,0 +1,8 @@
+# Settings
+
+# Deploy
+
+## When add new blog, project or change the slug of these
+
+- Rendering the sitemap data from generate-sitemap.xml and generate-sitemap-old.xml
+- Update static of the current host and old host url sitemap in public folder
