+++ pages/demo.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset>
+    <url>
+        <loc>Hello</loc>
+    </url>
+</urlset>
