+++ public/sitemap-old.xml
@@ -6,31 +6,31 @@
     <url>
         <loc>https://mal-blog.vercel.app/project</loc>
     </url>
-
+    <url>
+        <loc>https://mal-blog.vercel.app/blog</loc>
+    </url>
     <url>
         <loc>https://mal-blog.vercel.app/project/blog-mal</loc>
     </url>
     <url>
         <loc>https://mal-blog.vercel.app/project/chat-app</loc>
-    </url> 
-    <url>
-        <loc>https://mal-blog.vercel.app/blog</loc>
     </url>
-
+    <url>
+        <loc>https://mal-blog.vercel.app/blog/ecommerce-backend</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog.vercel.app/blog/miniapp-trello-clone</loc>
+    </url>
     <url>
         <loc>https://mal-blog.vercel.app/blog/new-year-2023</loc>
     </url>
-
     <url>
         <loc>https://mal-blog.vercel.app/blog/nextjs-and-mdx</loc>
     </url>
-
     <url>
         <loc>https://mal-blog.vercel.app/blog/react-input-box-with-textarea-and-contenteditable-p1</loc>
     </url>
-
     <url>
         <loc>https://mal-blog.vercel.app/blog/react-input-box-with-textarea-and-contenteditable-p2</loc>
     </url>
-
 </urlset>
