+++ public/sitemap.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/project</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/project/blog-mal</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/project/chat-app</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/ecommerce-backend</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/miniapp-trello-clone</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/new-year-2023</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/nextjs-and-mdx</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/react-input-box-with-textarea-and-contenteditable-p1</loc>
+    </url>
+    <url>
+        <loc>https://mal-blog-dawngate.vercel.app/blog/react-input-box-with-textarea-and-contenteditable-p2</loc>
+    </url>
+</urlset>
